+++ test/DGmethods/runtests.jl
@@ -23,7 +23,9 @@ using MPI, Test
                  (1, "Euler/isentropic_vortex_standalone_bc.jl")
                  (1, "conservation/sphere.jl")
                  (1, "compressible_Navier_Stokes/mms_bc.jl")
-                 (1, "sphere/advection_sphere.jl")
+                 (1, "sphere/advection_sphere_lsrk.jl")
+                 (1, "sphere/advection_sphere_ssp32.jl")
+                 (1, "sphere/advection_sphere_ssp43.jl")
                 ]
     cmd =  `mpiexec -n $n $(Base.julia_cmd()) --startup-file=no --project=$(Base.active_project()) --code-coverage=$coverage_opt $(joinpath(testdir, f))`
     @info "Running MPI test..." n f cmd
@@ -32,7 +34,6 @@ using MPI, Test
     # since external tests are not returned as passed/fail
     @test (run(cmd); true)
   end
-
 
   if "linux" != lowercase(get(ENV,"TRAVIS_OS_NAME",""))
     for (n, f) in [
